add 'primaryclient' module
This modules clears out the buffers whenever a client connects with
an identifier, making it essentially the "primary" client, and the
only one getting all the scrollback.

TODO: For now, this whole thing assumes there is only one client
that has an identifier. When there are more, we need to introduce
some sort of "sen_buffer_to_id = <identifier>" option, to specify
a specific identifier instead of just checking it its empty or not.